Vistas mejoradas - Requisitos de validación incluidos - Navegar entre si en cada vista

+++ database/migrations/2026_03_12_040204_create_ofertas_table.php
@@ -16,8 +16,8 @@ return new class extends Migration
             $table->string('titulo');
             $table->date('vigencia');
             $table->string('tienda');
-            $table->integer('precio_original');
-            $table->integer('precio_descuento');
+            $table->decimal('precio_original');
+            $table->decimal('precio_descuento');
             $table->timestamps();
         });
     }
